requirements/tests.txt
django_callable_perms
argparse
coverage
django-discover-runner
tox >= 1.8
